requirements.txt
django==4.2.11
uvloop==0.19.0
requests==2.31.0
ujson==5.9.0
aioredis==2.0.1
websockets==12.0
order-book==0.6.0
redis==5.0.3
msgpack==1.0.8
aiohttp==3.9.3
pytz==2024.1
django-redis==5.4.0
dj-redis-url==0.1.4
dj-database-url==2.1.0
aio-pika==9.4.1
pycryptodome==3.20.0
kombu==5.3.6
mysqlclient==2.2.4
PyMySQL==1.1.1